Tests/LibWeb/Text/expected/mathml-action-tabindex.txt

toggle.tabIndex=0
tooltip.tabIndex=-1
focused=true
